refactor(lifecycle): cerrar fase 4 (SDD en cliSdd, tipos en cli)
- Extrae comandos SDD a integrations/lifecycle/cliSdd.ts y carga perezosa desde cli.ts

- cli.ts importa solo tipos desde ../sdd (SddEvaluateResult, etc.)

- Actualiza plan-activo-de-trabajo: Fase 4 cerrada

Made-with: Cursor
